app/backend/pyproject.toml
[project]
name = "banking-assistant-copilot-backend"
version = "0.1.0"
description = "A multi-agent orchestration exposed as chat service for banking assistant frontend"
readme = "README.md"
requires-python = ">=3.11"

dependencies = [
    "fastapi[all]==0.116.1",
    "uvicorn==0.35.0",
    "httpx==0.28.1",
    "azure-identity==1.24.0",
    "azure-storage-blob==12.26.0",
    "azure-ai-documentintelligence==1.0.1",
    "dependency-injector==4.48.1",
    "agent-framework-azure-ai",
    "agent-framework-devui",
    "agent-framework-chatkit",
    "azure-monitor-opentelemetry-exporter==1.0.0b44"
]

[tool.uv]
prerelease = "if-necessary-or-explicit"
[tool.uv.sources]
"agent-framework-azure" = { git = "https://github.com/microsoft/agent-framework.git", branch = "main", subdirectory = "python/packages/azure-ai" }
"agent-framework-chatkit" = { git = "https://github.com/microsoft/agent-framework.git", branch = "main", subdirectory = "python/packages/chatkit" }
"agent-framework-core" = { git = "https://github.com/microsoft/agent-framework.git", branch = "main", subdirectory = "python/packages/core" }

[project.optional-dependencies]
dev = [
    "pytest==8.4.1",
    "pytest-asyncio==1.1.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"
